refactor: externalizes shortcut manager utility
Moves general shortcut handling code out of the KCodeBlock component into utilities. This way, other components can make use of it and leaves only the responsibility of providing a key map and a commands object to the components themselves.
